+++ test/e2e/outbound-form-tracker-test.js
@@ -23,6 +23,13 @@ import * as constants from '../../lib/constants';
 import pkg from '../../package.json';
 
 
+const DEFAULT_TRACKER_FIELDS = {
+  trackingId: 'UA-12345-1',
+  cookieDomain: 'auto',
+  siteSpeedSampleRate: 0,
+};
+
+
 const baseUrl = browser.options.baseUrl;
 
 
@@ -38,7 +45,7 @@ describe('outboundFormTracker', function() {
     log = bindLogAccessors(testId);
 
     browser.url('/test/e2e/fixtures/outbound-form-tracker.html');
-    browser.execute(ga.run, 'create', 'UA-XXXXX-Y', 'auto');
+    browser.execute(ga.run, 'create', DEFAULT_TRACKER_FIELDS);
     browser.execute(ga.logHitData, testId);
   });
 
@@ -87,7 +94,7 @@ describe('outboundFormTracker', function() {
 
     // Go back and submit a form that does match the `.form` selector.
     browser.url('/test/e2e/fixtures/outbound-form-tracker.html');
-    browser.execute(ga.run, 'create', 'UA-XXXXX-Y', 'auto');
+    browser.execute(ga.run, 'create', DEFAULT_TRACKER_FIELDS);
     browser.execute(ga.logHitData, testId);
     browser.execute(ga.run, 'require', 'outboundFormTracker', {
       formSelector: '.form',